backend/requirements.txt
# Django Core (4.2 LTS for Python 3.9+ compatibility, upgrade to 5.x when using Python 3.10+)
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# File handling
Pillow>=10.0.0
python-magic>=0.4.27

# Configuration
python-decouple>=3.8

# API Features
django-filter>=23.3
drf-spectacular>=0.27.0

# Rate limiting
django-ratelimit>=4.1.0

# Storage
django-storages>=1.14
boto3>=1.34.0

# Security
django-csp>=3.8
django-permissions-policy>=4.20.0

# Async tasks
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0

# AI/ML Integration
openai>=1.12.0

# Development & Testing
pytest>=8.0.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=22.0.0

# Code quality
ruff>=0.2.0
black>=24.0.0
isort>=5.13.0
bandit>=1.7.0
safety>=2.3.0
pip-audit>=2.7.0
